MovementController
I hope it's final version of controller
